% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidpub-package.R
\docType{package}
\name{covidpub-package}
\alias{covidpub}
\alias{covidpub-package}
\title{covidpub: COVID-19 in Connecticut}
\description{
DataHaven COVID-19 Analysis.
}
\author{
\strong{Maintainer}: Camille Seaberry \email{camille@ctdatahaven.org}

}
\keyword{internal}
